+++ Sources/Mapper.swift
@@ -17,61 +17,6 @@ public struct Mapper {
         self.JSON = JSON
     }
 
-    // MARK: - T
-
-    /**
-     Get a typed value from the given field in the source data
-
-     - parameter field: The field to retrieve from the source data, can be an empty string to return the
-                        entire data set
-
-     - throws: MapperError.MissingFieldError if the field doesn't exist
-     - throws: MapperError.TypeMismatchError if the value exists with the incorrect type
-
-     - returns: The value for the given field, if it can be converted to the expected type T
-     */
-    @warn_unused_result
-    public func from<T>(field: String) throws -> T {
-        let value = try self.JSONFromField(field)
-        if let value = value as? T {
-            return value
-        }
-
-        throw MapperError.TypeMismatchError(field: field, value: value, type: T.self)
-    }
-
-    /**
-     Get an optional typed value from the given field in the source data
-
-     - parameter field: The field to retrieve from the source data, can be an empty string to return the
-                        entire data set
-
-     - returns: The value for the given field, if it can be converted to the expected type T otherwise nil
-     */
-    @warn_unused_result
-    public func optionalFrom<T>(field: String) -> T? {
-        return try? self.from(field)
-    }
-
-    /**
-     Get an optional value from the given fields and source data. This returns the first non-nil value
-     produced in order based on the array of fields
-
-     - parameter fields: The array of fields to check from the source data.
-
-     - returns: The first non-nil value to be produced from the array of fields, or nil if none exist
-    */
-    @warn_unused_result
-    public func optionalFrom<T>(fields: [String]) -> T? {
-        for field in fields {
-            if let value: T = try? self.from(field) {
-                return value
-            }
-        }
-
-        return nil
-    }
-
     // MARK: - T: RawRepresentable
 
     /**
@@ -259,6 +204,25 @@ public struct Mapper {
     }
 
     /**
+     Get a Convertible value from a field in the source data
+
+     This transparently converts your types that conform to Convertible to properties on the Mappable type
+
+     - parameter field: The field to retrieve from the source data, can be an empty string to return the
+                        entire data set
+
+     - throws: Any error produced by the custom Convertible implementation
+
+     - note: This function is necessary because swift does not coerce the from that returns T to an optional
+
+     - returns: The value for the given field, if it can be converted to the expected type Optional<T>
+     */
+    @warn_unused_result
+    public func from<T: Convertible where T == T.ConvertedType>(field: String) throws -> T? {
+        return try self.from(field, transformation: T.fromMap)
+    }
+
+    /**
      Get an array of Convertible values from a field in the source data
 
      This transparently converts your types that conform to Convertible to an array on the Mappable type
@@ -309,6 +273,54 @@ public struct Mapper {
      */
     @warn_unused_result
     public func optionalFrom<T: Convertible where T == T.ConvertedType>(field: String) -> [T]? {
+        return try? self.from(field)
+    }
+
+    /**
+     Get a dictionary of Convertible values from a field in the source data
+
+     This transparently converts a source dictionary to a dictionary of 2 Convertible types
+
+     - parameter field: The field to retrieve from the source data, can be an empty string to return the
+                        entire data set
+
+     - throws: MapperError.TypeMismatchError if the value for the given field isn't a NSDictionary
+     - throws: Any error produced by the Convertible implementation of either expected type
+
+     - returns: A dictionary where the keys and values are created using their convertible implementations
+     */
+    @warn_unused_result
+    public func from<U: Convertible, T: Convertible
+        where U == U.ConvertedType, T == T.ConvertedType>(field: String) throws -> [U: T]
+    {
+        let object = try self.JSONFromField(field)
+        guard let data = object as? NSDictionary else {
+            throw MapperError.TypeMismatchError(field: field, value: object, type: NSDictionary.self)
+        }
+
+        var result = [U: T]()
+        for (key, value) in data {
+            result[try U.fromMap(key)] = try T.fromMap(value)
+        }
+
+        return result
+    }
+
+    /**
+     Get an optional dictionary of Convertible values from a field in the source data
+
+     This transparently converts a source dictionary to a dictionary of 2 Convertible types
+
+     - parameter field: The field to retrieve from the source data, can be an empty string to return the
+                        entire data set
+
+     - returns: A dictionary where the keys and values are created using their convertible implementations or
+                nil if anything throws
+     */
+    @warn_unused_result
+    public func optionalFrom<U: Convertible, T: Convertible
+        where U == U.ConvertedType, T == T.ConvertedType>(field: String) -> [U: T]?
+    {
         return try? self.from(field)
     }